requirements.txt
Django==1.9.6
MySQL-python==1.2.5
django-crispy-forms==1.6.1
python-dateutil
django-registration-redux==1.4
python-social-auth==0.2.12
beautifulsoup4
Pillow
django-active-link
django-extensions
pre-commit
black
flake8


